chore(size): swap toml for basic-toml to drop toml_edit (closes FU-A-06) (#22)
* chore(size): swap toml for basic-toml to drop toml_edit (closes FU-A-06)

Mirrors the workspace-side swap landed at kernex-dev/kernex@def75fc. Drops
toml_edit (~270 KiB on macOS) from the kernex-agent dep graph; cargo tree
-i toml and cargo tree -i toml_edit now both return "did not match any
packages".

Call sites migrated (3 from_str, 1 to_string_pretty):
- src/config.rs:63 toml::from_str -> basic_toml::from_str
- src/skills/manifest.rs:59 toml::from_str -> basic_toml::from_str
- src/skills/manifest.rs:67 toml::to_string_pretty -> basic_toml::to_string
  (compact; skills.toml is a machine-edited integrity file, the cosmetic
  diff is acceptable)
- src/serve/workflow.rs:37 toml::from_str -> basic_toml::from_str

Measured impact on macOS aarch64 release:
- Before: 15,779,248 bytes (15.05 MiB, ~49 KiB over the 15 MiB ceiling)
- After:  15,415,936 bytes (14.70 MiB, 312,704 bytes / ~305 KiB under)
- Delta:    -363,312 bytes (~355 KiB savings)

Closes FU-A-06. The macos-default-ceiling job in binary-size-build.yml
should flip green on this PR.

329 tests passing, clippy clean, fmt clean. The agent-codex feature
gate comment (line 80) updated to reflect that enabling agent-codex
now actually adds toml_edit (was previously a no-op size-wise because
toml_edit came in via toml transitively).

* fix(clippy): drop unnecessary i64 cast in parse_sqlite_utc

CI runs clippy under Rust 1.95.0 which surfaces unnecessary_cast on
`(days_since_epoch as i64)` (days_since_epoch is already i64 from
days_from_civil's return signature). Local toolchain didn't catch it.

Switched the hour/minute/second multiplications to `i64::from(...)`
(clippy's preferred form for u32 -> i64 lossless widening) while at it,
since they were the next-up cast warnings the lint would surface.

This entire `parse_sqlite_utc` / `days_from_civil` block deletes in
the memory-typed-row-shape Slice B paired-agent commit (B.6) when
the trait surfaces typed `updated_at: SystemTime` directly. Until
then, keep the block clippy-clean under Rust 1.95.

---------

+++ src/mem/cli.rs
@@ -506,11 +506,11 @@ fn parse_sqlite_utc(ts: &str) -> Option<SystemTime> {
     let second: u32 = ts.get(17..19)?.parse().ok()?;
 
     let days_since_epoch = days_from_civil(year, month, day)?;
-    let secs = (days_since_epoch as i64)
+    let secs = days_since_epoch
         .checked_mul(86_400)?
-        .checked_add(hour as i64 * 3_600)?
-        .checked_add(minute as i64 * 60)?
-        .checked_add(second as i64)?;
+        .checked_add(i64::from(hour) * 3_600)?
+        .checked_add(i64::from(minute) * 60)?
+        .checked_add(i64::from(second))?;
     if secs < 0 {
         return None;
     }
